remove GskIT
