+++ test-integration/server/session/BUILD
@@ -32,15 +32,18 @@ java_test(
 
 java_test(
     name = "transaction-oltp-it",
-    size = "medium",
+    size = "large",
     srcs = ["TransactionOLTPIT.java"],
     classpath_resources = ["//test-integration/resources:logback-test"],
     test_class = "grakn.core.server.session.TransactionOLTPIT",
     deps = [
+        "//api",
         "//common",
         "//concept",
+        "//dependencies/maven/artifacts/com/google/guava:guava",
         "//dependencies/maven/artifacts/org/apache/tinkerpop:gremlin-core",
         "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
+        "//dependencies/maven/artifacts/org/janusgraph:janusgraph-core",
         "//server",
         "//test-integration/rule:grakn-test-server",
         "//test-integration/util:graql-test-util",
